requirements.txt
django==4.2
chromadb==1.3.4
sentence-transformers==5.1.2
anthropic==0.73.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
firecrawl-py==4.8.0
